Reorg the fraud proof tests
Move the test of basic fraud proof features from cirrus-client-executor
to subspace-fraud-proof.
